Enable checkstyle

+++ java/src/main/java/io/cucumber/gherkin/utils/Accumulator.java
@@ -1,6 +1,6 @@
 package io.cucumber.gherkin.utils;
 
 public interface Accumulator {
-    void setDeepestLine(Long line);
-    Long getDeepestLine();
+    void setDeepestLine(Integer line);
+    Integer getDeepestLine();
 }
